rename: components/category-menu-components/CategoryLi.tsx → components/category-menu-components/CategoryMenuItems.tsx
